Fix int 32-bit overflow for consumer stats
For 32 bit systems, int size is 32, which leads to the broken stats.

+++ openvpn/middlewares/client/bytescount/fake_stats_handler.go
@@ -18,8 +18,8 @@
 package bytescount
 
 type fakeStatsRecorder struct {
-	BytesIn  int
-	BytesOut int
+	BytesIn  uint64
+	BytesOut uint64
 }
 
 func (sender *fakeStatsRecorder) record(bc Bytecount) error {
